+++ Word Bundle/Word Bundle/Views/Pages/RootView.swift
@@ -1,5 +1,5 @@
 //
-//  ZerothView.swift
+//  RootView.swift
 //  Word Bundle
 //
 //  Created by Giga Khizanishvili on 09.09.22.
@@ -7,7 +7,7 @@
 
 import SwiftUI
 
-struct ZerothView: View {
+struct RootView: View {
     @AppStorage(AppStorageKeys.hasCompletedOnboarding()) private var hasCompletedOnboarding = false
 
     var body: some View {
@@ -21,8 +21,8 @@ struct ZerothView: View {
     }
 }
 
-struct ZerothView_Previews: PreviewProvider {
+struct RootView_Previews: PreviewProvider {
     static var previews: some View {
-        ZerothView()
+        RootView()
     }
 }
